backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
Flask-Babel==4.0.0
PyJWT==2.8.0
pymongo[srv]==4.10.1
python-dotenv==1.0.0
cryptography==41.0.7
bcrypt==4.1.2
pytest==7.4.3
gunicorn==21.2.0
twilio==9.8.5
certifi==2024.8.30
reportlab==4.0.7
requests==2.31.0

google-generativeai>=0.8.0
